+++ packages/components/src/Tree/stories/FieldPicker.story.tsx
@@ -24,9 +24,8 @@
 
  */
 
-import React, { useState } from 'react'
+import React, { FC, useState } from 'react'
 import {
-  ButtonTransparent,
   IconButton,
   Menu,
   MenuDisclosure,
@@ -34,6 +33,7 @@ import {
   MenuItem,
   Tooltip,
   usePopover,
+  Badge,
 } from '../..'
 import { TreeItem, Tree, TreeGroup } from '..'
 
@@ -65,7 +65,7 @@ const PickerItem = ({ children = 'Cost', truncate = false }) => {
           detail={
             <>
               <IconButton
-                icon="Sync"
+                icon="Pivot"
                 label="Pivot"
                 tooltipPlacement="top"
                 onClick={(event) => {
@@ -112,50 +112,43 @@ const PickerItem = ({ children = 'Cost', truncate = false }) => {
   )
 }
 
-const addButton = (
-  <ButtonTransparent
-    size="xxsmall"
-    iconBefore="Plus"
-    onClick={() => alert('Hello Mouse')}
-    onKeyDown={(event) => {
-      if (event.key === 'Enter') {
-        event.preventDefault()
-        alert('Hello Keyboard')
-      }
-    }}
+const ViewTree: FC<{ children: string; defaultOpen?: boolean }> = ({
+  children,
+  defaultOpen,
+}) => (
+  <Tree
+    defaultOpen={defaultOpen}
+    detail={<Badge intent="inform">1</Badge>}
+    detailAccessory
+    indicatorIcons={{ close: 'CaretRight', open: 'CaretDown' }}
+    indicatorPosition="right"
+    label={children}
   >
-    Add
-  </ButtonTransparent>
+    <TreeGroup label="DIMENSIONS">
+      <Tree branchFontWeight label="Created">
+        <PickerItem>Created Date</PickerItem>
+        <PickerItem>Created Month</PickerItem>
+        <PickerItem>Created Year</PickerItem>
+      </Tree>
+      <PickerItem>City</PickerItem>
+      <PickerItem>Country</PickerItem>
+      <PickerItem>ID</PickerItem>
+    </TreeGroup>
+    <TreeGroup label="MEASURES" color="orange">
+      <Tree branchFontWeight label="My Measure Group">
+        <PickerItem>Count of IDs</PickerItem>
+        <PickerItem>Count of Cities</PickerItem>
+      </Tree>
+      <PickerItem>Sum</PickerItem>
+      <PickerItem>Max</PickerItem>
+    </TreeGroup>
+  </Tree>
 )
 
 export const FieldPicker = () => (
-  <Tree defaultOpen detailAccessory detail={addButton} label="Custom Fields">
-    <TreeGroup label="DIMENSIONS">
-      <PickerItem />
-      <PickerItem />
-      <PickerItem />
-      <PickerItem>
-        Over a thousand types of cheese exist and are currently produced in
-        various countries. Their styles, textures and flavors depend on the
-        origin of the milk (including the animal's diet), whether they have been
-        pasteurized, the butterfat content, the bacteria and mold, the
-        processing, and how long they have been aged for.
-      </PickerItem>
-      <PickerItem truncate>
-        Herbs, spices, or wood smoke may be used as flavoring agents. The yellow
-        to red color of many cheeses is produced by adding annatto. Other
-        ingredients may be added to some cheeses, such as black pepper, garlic,
-        chives or cranberries.
-      </PickerItem>
-    </TreeGroup>
-    <TreeGroup label="MEASURES" color="orange">
-      <Tree branchAlign branchFontWeight label="Hello">
-        <PickerItem />
-      </Tree>
-      <TreeItem>Name</TreeItem>
-      <PickerItem />
-      <PickerItem />
-      <PickerItem />
-    </TreeGroup>
-  </Tree>
+  <>
+    <ViewTree defaultOpen={true}>Orders</ViewTree>
+    <ViewTree>Orders Items</ViewTree>
+    <ViewTree>Users</ViewTree>
+  </>
 )